added task list
To get started, just tap or click this placeholder text and begin typing. You can view and edit this document on your Mac, iPad, iPhone, or on iCloud.com.
It’s easy to edit text, change fonts, and add beautiful graphics. Use paragraph styles to get a consistent look throughout your document. For example, this paragraph uses Body style. You can change it in the Text tab of the Format controls.
To add photos, movies, audio, and other objects, tap or click one of the buttons in the toolbar or drag and drop the objects onto the page.
You can use Pages for both word processing and page layout. This Essay template is set up for word processing, so your text flows from one page to the next as you type, with new pages created automatically when you reach the end of a page. 
In page layout documents, you can manually rearrange pages and freely position text boxes, images, and other objects on the page. To create a page layout document, choose a page layout template in the template chooser. You can also change this document to page layout on your Mac, iPad, or iPhone by turning off Document Body in the Document controls.